deprecate the grinder entrypoint
