+++ bin/grind.dart
@@ -6,6 +6,35 @@
  */
 library grinder.grind;
 
-import 'grinder.dart' as g;
+import 'dart:io';
 
-void main(List args) => g.runScript('tool/grind.dart', args);
+void main(List args) => runScript('tool/grind.dart', args);
+
+void runScript(String script, List args) {
+  File file = new File(script);
+
+  if (!file.existsSync()) {
+    print("Error: expected to find '${script}' "
+        "relative to the current directory.");
+    exit(1);
+  }
+
+  List newArgs = [script]..addAll(args);
+  _runProcessAsync(Platform.isWindows ? 'dart.exe' : 'dart', newArgs);
+}
+
+void _runProcessAsync(String executable, List<String> arguments) {
+  Process.start(executable, arguments).then((Process process) {
+    process.stdout.listen((List<int> data) {
+      stdout.write(new String.fromCharCodes(data));
+    });
+
+    process.stderr.listen((List<int> data) {
+      stderr.write(new String.fromCharCodes(data));
+    });
+
+    return process.exitCode.then((int code) {
+      if (code != 0) exit(code);
+    });
+  });
+}
