+++ bin/grinder.dart
@@ -2,38 +2,16 @@
 // governed by a BSD-style license that can be found in the LICENSE file.
 
 /**
- * Look for `tool/grinder.dart` relative to the current directory and run it.
+ * Look for `tool/grind.dart` relative to the current directory and run it.
  */
 library grinder.grinder;
 
-import 'dart:io';
+import 'grind.dart' as grind;
 
-void main(List args) => runScript('tool/grinder.dart', args);
+@deprecated
+void main(List args) {
+  print("The 'grinder' entrypoint is deprecated; please use the 'grind'.");
+  print('');
 
-void runScript(String script, List args) {
-  File file = new File(script);
-
-  if (!file.existsSync()) {
-    print("Error: expected to find a '${script}' script.");
-    exit(1);
-  }
-
-  List newArgs = [script]..addAll(args);
-  _runProcessAsync(Platform.isWindows ? 'dart.exe' : 'dart', newArgs);
-}
-
-void _runProcessAsync(String executable, List<String> arguments) {
-  Process.start(executable, arguments).then((Process process) {
-    process.stdout.listen((List<int> data) {
-      stdout.write(new String.fromCharCodes(data));
-    });
-
-    process.stderr.listen((List<int> data) {
-      stderr.write(new String.fromCharCodes(data));
-    });
-
-    return process.exitCode.then((int code) {
-      if (code != 0) exit(code);
-    });
-  });
+  grind.main(args);
 }
