+++ changelog.md
@@ -1,5 +1,8 @@
 # grinder.dart changes
 
+## 0.6.4
+- deprecate the `grinder` entrypoint; users should start the tool with `grind`
+
 ## 0.6.2 (2014/11/13)
 
 - widen the version constraint on `quiver`
@@ -12,16 +15,17 @@
 
 The convenience API is now more terse:
 - the `defineTask()` method has been renamed to `task()`
-- instead of named parameters, the `task()` function now uses optional positional parameters
+- instead of named parameters, the `task()` function now uses optional
+  positional parameters
 
 Added two new entrypoint files in `bin/`: `grind.dart` and `grinder.dart`. These
 let you run grinder via:
 
     pub run grinder test
 
-They look for a cooresponding grinder script in the `tool` directory (`bin/grind.dart`
+They look for a corresponding grinder script in the `tool` directory (`bin/grind.dart`
 looks for `tool/grind.dart` and `bin/grinder.dart` looks for `tool/grinder.dart`).
-If they find a cooresponding script they run it in a new Dart VM process. This
+If they find a corresponding script they run it in a new Dart VM process. This
 means that projects will no longer have to have a `grind.sh` script in the root
 of each project.
 
@@ -37,6 +41,8 @@ dartanalyzer - `Analyzer` - was created.
 
 ## 0.5.7 (2014/07/28)
 
-- added `runProcessAsync()` and related async methods (such as `PubTools.buildAsync(...)`)
+- added `runProcessAsync()` and related async methods (such as 
+  `PubTools.buildAsync(...)`)
 - removed duplicated stack traces when the build fails with exceptions
-- throw an exception when running SDK binaries, and we are not able to locate the Dart SDK
+- throw an exception when running SDK binaries, and we are not able to locate
+  the Dart SDK
